#26 - refactor:hashTag위치변경
