--- Dockerfile
@@ -1,5 +1,0 @@
-FROM openjdk:21-jdk
-
-COPY build/libs/*.jar app.jar
-
-ENTRYPOINT ["java","-jar","/app.jar"]
