+++ src/main/java/com/example/peep/domain/enumType/HashtagType.java
@@ -1,4 +1,4 @@
-package com.example.peep.enumType;
+package com.example.peep.domain.enumType;
 
 public enum HashtagType {
     HOBBY,
