+++ src/main/java/com/example/peep/repository/HashtagRepository.java
@@ -1,7 +1,7 @@
 package com.example.peep.repository;
 
 import com.example.peep.domain.Hashtag;
-import com.example.peep.enumType.HashtagType;
+import com.example.peep.domain.enumType.HashtagType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
